Workaround bug in npm v5+
When a project uses a package-lock.json, it deletes unlisted dependencies
before the `install` script is ran (though it doesn't much matter when).
This invalidates the cache file for `npm-svn` and causes errors since
`npm-svn` thinks the SVN-based dependency is installed.

This workaround "fixes" the issue by ignoring the cache -- always -- when
checking to see if a project needs to be updated. This, in turn, causes
dependencies to always be reinstalled.

This is not the ideal fix, but is sufficient for now. This branch will
(hopefully) eventually contain the correct fix and is intended as a band-aid.
This branch should never contain breaking changes and is intended to only be
used so long as the fix hasn't made into the published version on npm.